LibWeb: Extract AudioParam's automation timeline into its own class
The automation event list and its evaluation logic are self-contained,
and upcoming audio rendering work needs to evaluate them outside of the
GC-managed AudioParam object. Move them into a ref-counted
AudioParamTimeline that AudioParam delegates to. No functional changes.
